requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cleanup==5.2.0
django-environ==0.7.0
django-filter==21.1
django-resized==0.3.11
django-widget-tweaks==1.4.8
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
